requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
# created-by: conda 26.1.1
_openmp_mutex=4.5=20_gnu
aiofiles=25.1.0=pypi_0
aiosqlite=0.22.1=pypi_0
apscheduler=3.11.2=pypi_0
asyncua=1.1.8=pypi_0
black=26.3.1=pypi_0
bzip2=1.0.8=hda65f42_9
c104=2.2.1=pypi_0
ca-certificates=2026.4.22=hbd8a1cb_0
cffi=2.0.0=pypi_0
click=8.3.2=pypi_0
cryptography=46.0.7=pypi_0
fakeredis=2.35.1=pypi_0
flake8=7.3.0=pypi_0
flake8-docstrings=1.7.0=pypi_0
giflib=5.2.2=hd590300_0
greenlet=3.4.0=pypi_0
icu=78.3=h33c6efd_0
iniconfig=2.3.0=pypi_0
isort=8.0.1=pypi_0
kafka-python=2.3.1=pypi_0
kafka-python-ng=2.2.3=pypi_0
ld_impl_linux-64=2.45.1=default_hbd61a6d_102
leptonica=1.87.0=h12c84f9_1
lerc=4.1.0=hdb68285_0
libarchive=3.8.7=gpl_hc2c16d8_100
libdeflate=1.25=h17f619e_0
libexpat=2.7.5=hecca717_0
libffi=3.5.2=h3435931_0
libgcc=15.2.0=he0feb66_18
libgcc-ng=15.2.0=h69a702a_18
libgomp=15.2.0=he0feb66_18
libiconv=1.18=h3b78370_2
libjpeg-turbo=3.1.4.1=hb03c661_0
liblzma=5.8.3=hb03c661_0
libnsl=2.0.1=hb9d3cd8_1
libpng=1.6.58=h421ea60_0
librt=0.9.0=pypi_0
libsqlite=3.53.0=hf4e2dac_0
libstdcxx=15.2.0=h934c35e_18
libtiff=4.7.1=h9d88235_1
libuuid=2.42=h5347b49_0
libuv=1.51.0=hb03c661_1
libwebp-base=1.6.0=hd42ef1d_0
libxcrypt=4.4.36=hd590300_1
libxml2=2.15.3=h49c6c72_0
libxml2-16=2.15.3=hca6bf5a_0
libzlib=1.3.2=h25fd6f3_2
lz4-c=1.10.0=h5888daf_1
lzo=2.10=h280c20c_1002
mccabe=0.7.0=pypi_0
mypy=1.20.1=pypi_0
mypy-extensions=1.1.0=pypi_0
ncurses=6.5=h2d0b736_3
nodejs=22.22.2=h273caaf_0
openjpeg=2.5.4=h55fea9a_0
openssl=3.6.2=h35e630c_0
packaging=26.0=pyhcf101f3_0
pathspec=1.0.4=pypi_0
pillow=12.2.0=pypi_0
pip=26.0.1=pyh8b19718_0
platformdirs=4.9.6=pypi_0
pluggy=1.6.0=pypi_0
psycopg=3.3.3=pypi_0
psycopg-binary=3.3.3=pypi_0
pycodestyle=2.14.0=pypi_0
pycparser=3.0=pypi_0
pydocstyle=6.3.0=pypi_0
pyflakes=3.4.0=pypi_0
pygments=2.20.0=pypi_0
pymupdf=1.27.2.3=pypi_0
pyopenssl=26.0.0=pypi_0
pytesseract=0.3.13=pypi_0
pytest=9.0.3=pypi_0
python=3.11.15=hd63d673_0_cpython
python-dateutil=2.9.0.post0=pypi_0
pytokens=0.4.1=pypi_0
pytz=2026.1.post1=pypi_0
pyyaml=6.0.3=pypi_0
readline=8.3=h853b02a_0
redis=7.4.0=pypi_0
setuptools=82.0.1=pyh332efcf_0
six=1.17.0=pypi_0
snowballstemmer=3.0.1=pypi_0
sortedcontainers=2.4.0=pypi_0
sqlalchemy=2.0.49=pypi_0
tesseract=5.5.2=h1f36d38_0
tk=8.6.13=noxft_h366c992_103
types-pyyaml=6.0.12.20260408=pypi_0
typing-extensions=4.15.0=pypi_0
tzdata=2025c=hc9c84f9_1
tzlocal=5.3.1=pypi_0
wait-for2=0.3.2=pypi_0
wheel=0.46.3=pyhd8ed1ab_0
zstd=1.5.7=hb78ec9c_6
